review(routing): address Claude, Codex, CodeRabbit round-1 findings
Aggregated 9 actionable findings from three independent reviewers
(local Claude subagent, codex CLI, GitHub CodeRabbit) and applied
the must-fix set:

Security / correctness
- Mix server-side HMAC salt (MAXX_ROUTING_SEED_SALT) into the seeded
  shuffle so an authenticated client can no longer grind X-Session-Id
  values to steer their own traffic to a specific upstream (Codex+Claude).
- Anchor the seed to the routing context (tenant/client/project) so
  unauthenticated setups (apiTokenID=0) get deterministic distribution
  instead of falling back to time.Now() and clobbering Phase B's
  determinism contract (Claude).
- Bump policyFingerprint from 48 bits to 128 bits. The earlier birthday
  bound (~1.6e7 distinct configs) was a long-tail adversarial concern;
  128 bits removes it without making Redis keys meaningfully longer
  (Codex+Claude).
- Sticky write now uses a fresh 2s background context instead of
  state.ctx, which can already be Done by the time the dispatcher
  reaches the write path on streaming responses — every SET was
  silently failing under load. Read path was already using
  context.Background; this restores symmetry (Claude).
- Bound the sticky read (Match) with a 100ms WithTimeout so a slow
  or unavailable Redis can't stall the route match (CodeRabbit).

Robustness
- memoryStore.Get lazily deletes expired entries under a write-lock
  re-check. Redis backend gets free TTL eviction; the in-process
  store was accumulating expired records forever (CodeRabbit).
- StickyTTLSeconds is now int64, removing the 32-bit overflow risk
  for very long TTLs (Claude).
- Document the SetCoordinator lifecycle in sticky.Manager: the atomic
  pointer swap is safe under concurrent reads, but in-flight calls
  against the old store keep running, so callers shouldn't tear that
  store down until after the swap (Codex+Claude).

Tests
- Refactored TestVerifyRoutingErrorClass to use symmetric primed/spare
  side roles instead of two asymmetric branches (the primed=B branch
  was using srvB.CloseClientConnections() which the reviewer flagged
  as unreliable — listener stays open). Both initial pinnings now drive
  the same setStatus(500) failure injection (Codex+CodeRabbit).
- Added an explicit 429 sub-probe: with sticky already pinned, flip
  the pinned side to 429 and verify the failover target is reached
  and sticky overwrites correctly. The dispatcher doesn't branch by
  error class today, but Codex review wanted observable 429 coverage
  to catch any future rate-limit-specific drift (Codex).
- store_redis_test.go now asserts err == nil on Get calls instead of
  discarding the error with _, so a real Redis failure surfaces
  immediately rather than masquerading as "not found" (CodeRabbit).
- Strengthen the Phase D cooldown probe: previously only asserted
  the cooled-down provider got 0 hits; now also bounds the surviving
  providers' counts to ±20 of expected (133/67 for 200 sessions
  with weights 2:1), so weighted failover regressions get caught
  numerically rather than logged-but-passing (Claude).

False positives examined and not applied
- Claude/CodeRabbit both flagged Phase D cooldown cleanup as
  incomplete (only one of three keys cleared). cooldown Manager
  treats ClearCooldown(providerID, "", "") as "clear all variants
  for this provider" (manager.go:402-413), so the existing cleanup
  is correct; the explicit per-variant clears would be redundant.
- Claude flagged Manager.SetStore as racy with in-flight Get/Set.
  The atomic.Pointer pattern is intentional: in-flight calls
  finish against the old backend, new calls hit the new one.
  Documented the constraint instead of changing the design.

Remaining items reviewed and deferred (not blockers): test
table-driven refactor (Codex NIT, would lose log readability),
hot-key documentation note (Codex NIT, marginal under LLM
latencies), GetByID O(N) cache scan (low call rate from admin
UI), frontend grid responsive minor (cosmetic).

All tests pass including the strengthened probes and the new
429 path; full e2e + multi-instance suites green.

+++ internal/domain/model.go
@@ -629,9 +629,9 @@ type RoutingStrategyConfig struct {
 	// Sticky / session-affinity 配置（用于 weighted_random 策略；priority 下忽略）
 	// 启用后：同一 (api token[+session]) 命中过的 provider 会在 TTL 内被记住，
 	// 后续请求优先尝试它（最大化上游 prompt cache 命中率）。
-	StickyEnabled    bool                 `json:"stickyEnabled,omitempty"`
-	StickyScope      RoutingStickyScope   `json:"stickyScope,omitempty"`      // "token" | "conversation"，默认 "token"
-	StickyTTLSeconds int                  `json:"stickyTTLSeconds,omitempty"` // 默认 1800（30 分钟），<=0 取默认
+	StickyEnabled    bool               `json:"stickyEnabled,omitempty"`
+	StickyScope      RoutingStickyScope `json:"stickyScope,omitempty"`      // "token" | "conversation"，默认 "token"
+	StickyTTLSeconds int64              `json:"stickyTTLSeconds,omitempty"` // 默认 1800（30 分钟），<=0 取默认
 }
 
 type RoutingStickyScope string
